refactor: 페이지 사이드바 FSD 아키텍처로 변경
