+++ apps/frontend/src/features/canvas/ui/Node/NoteNode/index.tsx
@@ -2,10 +2,10 @@ import { useEffect, useState } from "react";
 import { Handle, NodeProps, Position, type Node } from "@xyflow/react";
 import { FileText } from "lucide-react";
 
-import usePageStore from "@/store/usePageStore";
 import useUserStore from "@/store/useUserStore";
 import Emoji from "@/components/commons/emoji";
 import ActiveUser from "@/components/commons/activeUser";
+import { usePageStore } from "@/features/pageSidebar/model";
 
 export type NoteNodeData = { title: string; id: number; emoji: string };
 export type NoteNodeType = Node<NoteNodeData, "note">;
