+++ apps/frontend/src/features/pageSidebar/ui/index.ts
@@ -1,0 +1,7 @@
+export * from "./LogoBtn";
+export * from "./NoteList";
+export * from "./ProfileModal";
+export * from "./RemoveNoteModal";
+export * from "./Tools";
+export * from "./TopNav";
+export * from "./WorkspaceNav";
